attempt to use latest ps2dev with 1400 mcman and padman

+++ iop/ps2host/net_fio.c
@@ -420,7 +420,7 @@ int pko_read_file(int fd, char *buf, int length)
 
 //----------------------------------------------------------------------
 //
-int pko_ioctl(int fd, unsigned long request, void *data)
+int pko_ioctl(int fd, int request, void *data)
 {
 	pko_pkt_ioctl_req *ioctlreq;
 	pko_pkt_file_rly *ioctlrly;
